Move backend to k8s-specific package
Create API definition
